ti_Tag should always be treated as ULONG
git-svn-id: https://svn.aros.org/svn/aros/trunk@51766 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/arossupport/libnexttagitem.c
@@ -12,7 +12,7 @@ struct TagItem *LibNextTagItem(struct TagItem **tagListPtr)
 
     while(1)
     {
-        switch((*tagListPtr)->ti_Tag)
+        switch((ULONG)((*tagListPtr)->ti_Tag))
         {
             case TAG_MORE:
                 if (!((*tagListPtr) = (struct TagItem *)(*tagListPtr)->ti_Data))
